refactor(reservations): deepen mutation lifecycle
